# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TasteTwin API",
    "description": "Nutrition-aware meal recommendations and Qloo cultural taste discovery. Menu provenance is separate from place affinity.",
    "version": "0.2.0"
  },
  "paths": {
    "/api/me": {
      "get": {
        "summary": "My Profile",
        "operationId": "my_profile_api_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "summary": "Update My Profile",
        "operationId": "update_my_profile_api_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/places": {
      "post": {
        "summary": "Discover Places",
        "operationId": "discover_places_api_places_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlacesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/taste/tags": {
      "get": {
        "summary": "Taste Tags",
        "operationId": "taste_tags_api_taste_tags_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80,
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interpret": {
      "post": {
        "summary": "Parse",
        "operationId": "parse_api_interpret_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterpretRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/recommend": {
      "post": {
        "summary": "Recommend",
        "operationId": "recommend_api_recommend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/refine": {
      "post": {
        "summary": "Refine",
        "operationId": "refine_api_refine_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/select": {
      "post": {
        "summary": "Select",
        "operationId": "select_api_select_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/sessions/{session_id}": {
      "delete": {
        "summary": "Delete Session",
        "operationId": "delete_session_api_sessions__session_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Session Records",
        "operationId": "session_records_api_sessions__session_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlacesRequest": {
        "properties": {
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "tasteProfile": {
            "$ref": "#/components/schemas/TasteProfile"
          },
          "radiusMeters": {
            "type": "integer",
            "maximum": 20000,
            "minimum": 100,
            "title": "Radiusmeters",
            "default": 2000
          },
          "limit": {
            "type": "integer",
            "maximum": 50,
            "minimum": 1,
            "title": "Limit",
            "default": 10
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "location"
        ],
        "title": "PlacesRequest"
      },
      "ProfileUpdate": {
        "properties": {
          "nickname": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Nickname"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other",
              "undisclosed"
            ],
            "title": "Gender"
          },
          "birthYear": {
            "type": "integer",
            "maximum": 9999,
            "minimum": 1,
            "title": "Birthyear"
          },
          "birthMonth": {
            "type": "integer",
            "maximum": 12,
            "minimum": 1,
            "title": "Birthmonth"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nickname",
          "gender",
          "birthYear",
          "birthMonth"
        ],
        "title": "ProfileUpdate"
      },
      "RecommendRequest": {
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid",
            "title": "Sessionid"
          },
          "qlooEnabled": {
            "type": "boolean",
            "title": "Qlooenabled",
            "default": true
          },
          "limit": {
            "type": "integer",
            "maximum": 3,
            "minimum": 1,
            "title": "Limit",
            "default": 3
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sessionId"
        ],
        "title": "RecommendRequest"
      },
      "RefineRequest": {
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid",
            "title": "Sessionid"
          },
          "refinement": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Refinement"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sessionId",
          "refinement"
        ],
        "title": "RefineRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}